Replace console.log with log4js
